fix(agent-runs): store blocker ids as bigint array

+++ db/migrate/20260518053300_add_blocked_by_issue_ids_to_agent_runs.rb
@@ -2,7 +2,7 @@
 
 class AddBlockedByIssueIdsToAgentRuns < ActiveRecord::Migration[8.1]
   def change
-    add_column :agent_runs, :blocked_by_issue_ids, :integer, array: true, default: [],
+    add_column :agent_runs, :blocked_by_issue_ids, :bigint, array: true, default: [],
       comment: "IDs of issues/PRs that block the created issue from being picked up for work."
   end
 end
